hab/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.6
click==8.0.1
colorama==0.4.4
Django==3.2.8
django-filter==21.1
djangorestframework==3.12.4
idna==3.2
Markdown==3.3.4
PyQt5==5.15.4
pyqt5-plugins==5.15.4.2.2
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyqt5-tools==5.15.4.3.2
python-dotenv==0.19.0
pytz==2021.3
qt5-applications==5.15.2.2.2
qt5-tools==5.15.2.1.2
requests==2.26.0
sqlparse==0.4.2
urllib3==1.26.7
